Move the filling logic to calendar.tsx.
